Changed the return type of complete+start task and delete log
Completing a task will now also return the previous active log if
exist.

Starting a log will return the active log. This was done so that
the user who made the request could get the logID with it.

Deleting a log will now return the new total time of the log. This
was done so that the user may know the affects of deleting the log.